Documentation: Update README and license
License is GPLv2; currently, no relicensing to newer is allowed because
of clause 9. This should stay unless really good arguments are made
otherwise.

Additionally, update the readme. There's a lot of overlap between the
'-h' and README.md files, but I'm okay with that until someone
volunteers to do a better job.

Signed-off-by: Aaron Conole <aconole@redhat.com>