SearchMailman.py: Add the ability to extract individual messages
It is sometimes very useful to have the ability to filter out specific
messages and save them off individually (this is extra important when using
development mailing lists and trying to obtain patchsets). This commit
takes the argument given to -o and detects if it is a directory. If so, it
assumes the user wants individual mbox files, and operates accordingly.

Signed-off-by: Aaron Conole <aconole@redhat.com>