SearchMailman.py: Detect patches
When saving individual files, this detects that a patch is part of a series
and outputs it with the patchnum as the mailnum ID instead of the order it
was encountered in the mailing list. This is useful for applying a whole
series of patches extracted with a SearchMailman query.

Signed-off-by: Aaron Conole <aconole@redhat.com>